requirements.txt
python-telegram-bot[job-queue]==21.9
openai>=1.40.0
aiohttp>=3.9.0
RestrictedPython>=7.0
croniter>=2.0.0
